requirements.txt
apache-airflow
apache-airflow-providers-cncf-kubernetes
apache-airflow-providers-apache-spark
kubernetes
boto3